backend/requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.4.2
google-generativeai==0.8.3
